Add new implementations and strcuture
