+++ benchmarks/gol.py
