+++ build_scripts/collision.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+import os, sys
+currentdir = os.path.dirname(os.path.realpath(__file__))
+parentdir = os.path.dirname(currentdir)
+sys.path.append(parentdir + '/src')
+
+from sanajeh import PyCompiler
+
+compiler: PyCompiler = PyCompiler("examples/collision/collision.py", "collision")
+compiler.compile()
+compiler.build()
