--- device_code/nested/nested.cdef
@@ -1,8 +1,0 @@
-
-int Body_Body_compute_force();
-int Body_Body_body_update();
-int parallel_new_Body(int object_num);
-int parallel_new_Vector(int object_num);
-int AllocatorInitialize();
-int Vector_do_all(void (*pf)());
-int AllocatorUninitialize();
