--- memo
@@ -1,113 +1,0 @@
-import先の関数の扱い
-
-# Sanajehの制限
-1 python3.7で書く
-    ->関数の返り値の型の宣言が必要(ない時はvoidと認識される)
-    ->クラスのfieldとその型の宣言が必要
-2 parallel_do can only use functions with specific syntax (for now, "parallel_do" itself) which has been specified in device code
-3 for nested objects, fields must be declared in "_init__"
-
-
- ------use eval-------
-(100000 objects, 100 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.310s
-initialize time: 0.115s
-parallel new time(100000 objects): 0.014s
-average computation time: 0.413s
-overall computation time(100  iterations): 41.332s
-do_all time: 8.182s
-overall time: 55.959s
-
-(100000 objects, 10 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.196s
-initialize time: 0.117s
-parallel new time(100000 objects): 0.014s
-average computation time: 0.417s
-overall computation time(10   iterations): 4.174s
-do_all time: 8.343s
-overall time: 18.851s
-
-(10000 objects, 10 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.217s
-initialize time: 0.112s
-parallel new time(10000 objects): 0.005s
-average computation time: 0.013s
-overall computation time(10   iterations): 0.131s
-do_all time: 0.718s
-overall time: 7.189s
-
-(10000 objects, 100 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.263s
-initialize time: 0.128s
-parallel new time(10000 objects): 0.005s
-average computation time: 0.011s
-overall computation time(100  iterations): 1.085s
-do_all time: 0.425s
-overall time: 7.913s
-
-(1000 objects, 10 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.282s
-initialize time: 0.125s
-parallel new time(1000  objects): 0.004s
-average computation time: 0.001s
-overall computation time(10   iterations): 0.015s
-do_all time: 0.040s
-overall time: 6.472s
-
------without eval -------
-(100000 objects, 100 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.273s
-initialize time: 0.124s
-parallel new time(100000 objects): 0.014s
-average computation time: 0.418s
-overall computation time(100  iterations): 41.755s
-do_all time: 7.551s
-overall time: 55.723s
-
-(100000 objects, 10 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.218s
-initialize time: 0.125s
-parallel new time(100000 objects): 0.013s
-average computation time: 0.416s
-overall computation time(10   iterations): 4.163s
-do_all time: 8.085s
-overall time: 18.610s
-
-
-(10000 objects, 10 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.241s
-initialize time: 0.119s
-parallel new time(10000 objects): 0.005s
-average computation time: 0.013s
-overall computation time(10   iterations): 0.131s
-do_all time: 0.759s
-overall time: 7.262s
-
-(10000 objects, 100 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.246s
-initialize time: 0.120s
-parallel new time(10000 objects): 0.006s
-average computation time: 0.011s
-overall computation time(100  iterations): 1.079s
-do_all time: 0.731s
-overall time: 8.187s
-
-(1000 objects, 10 iterations)
-compile time(py2cpp): 0.006s
-compile time(nvcc): 6.292s
-initialize time: 0.108s
-parallel new time(1000  objects): 0.004s
-average computation time: 0.001s
-overall computation time(10   iterations): 0.011s
-do_all time: 0.037s
-overall time: 6.457s
-
